NVISH_exercise/requirements.txt
alembic==1.11.3
async-timeout==4.0.3
blinker==1.6.2
click==8.1.7
exceptiongroup==1.1.3
Flask==2.3.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
python-dotenv==1.0.0
redis==5.0.0
SQLAlchemy==2.0.20
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.7
